requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
PyYAML>=6.0.1

# --- for FastAPI deployment ---
fastapi
uvicorn[standard]
pydantic
